edit.js file modified
