+++ src/index.html
@@ -5,31 +5,20 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>To-Do List</title>
+    <link rel="stylesheet" href="styles.css">
 </head>
 <body>
     <div id="app">
-      <h1>Today's To Do</h1>
-      <div class="addTask">
-        <input type="text" name="" class="addTask" id="inputTask" placeholder="Add a task...">
-        <button id="button">Add</button>
-      </div>
-      <!-- <ul id="todo-list"></ul> -->
-      <div class="container">
-        <div class="content">
-          <div class="toDoList">
-            <input type="checkbox" name="" id="">
-            <input type="text" name="" id="" value="text">
-          </div>
-          <i class="fa-regular fa-ellipsis-vertical"></i>
+        <h1>Today's To Do</h1>
+        <div class="addTask">
+            <input type="text" name="" class="addTask" id="inputTask" placeholder="Add a task...">
+            <button id="button">Add</button>
         </div>
-      </div>
-      <div class="button-container">
-        <button id="clearButton" class="clear">Clear as Complete</button>
-      </div>
+        <ul id="todo-list"></ul>
+        <div class="button-container">
+            <button id="clearButton" class="clear">Clear as Complete</button>
+        </div>
     </div>
-    <script src="index.js"></script>
-  </body>
-  
+    <script src="dist/bundle.js"></script>
+</body>
 </html>
-
-
