+++ webpack.config.js
@@ -3,10 +3,10 @@ const path = require('path');
 
 module.exports = {
   mode: 'development',
-  entry: '/src/index.js',
-
+  entry: './src/index.js',
   output: {
     path: path.resolve(__dirname, 'dist'),
+    filename: 'bundle.js',
   },
   module: {
     rules: [
@@ -18,7 +18,7 @@ module.exports = {
   },
   plugins: [
     new HtmlWebpackPlugin({
-      template: path.resolve(__dirname, 'src', 'index.html'),
+      template: './src/index.html',
     }),
   ],
 };
